backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==23.2.1
ruff==0.8.4
